requirements.txt
fastapi>=0.111.0,<0.112.0
uvicorn[standard]>=0.29.0,<0.30.0
pydantic>=2.7.1,<3.0.0
python-dotenv>=1.0.1,<2.0.0
google-generativeai>=0.5.4,<0.6.0
reportlab>=4.2.0,<5.0.0
python-dotenv
